requirements.txt
pygame==2.5.2
requests==2.31.0
beautifulsoup4==4.12.3
Pillow==10.2.0
